+++ doc/whats_new.rst
@@ -69,6 +69,8 @@ Changelog
 
     - Add :func:`mne.Epochs.load_data` method to :class:`mne.Epochs` by `Teon Brooks`_
 
+    - Add support for drawing scalp maps by selecting an area in :func:`mne.Evoked.plot` by `Jaakko Leppakangas`_
+
 BUG
 ~~~